+++ apps/userprofile/forms.py
@@ -12,17 +12,20 @@ from apps.userprofile.models import UserProfile
 class UserProfileForm(forms.ModelForm):
     class Meta:
         model = UserProfile
-        fields = ('nick', 'date_of_birth', 'address', 'zip_code', 'phone')
+        fields = ('nick', 'date_of_birth', 'address', 'zip_code', 'phone', 'marketing_optin')
         widgets = {
             'nick': forms.TextInput(attrs={'class': 'form-control', 'placeholder': _(u'Nickname'), 'type': 'text'}),
             'date_of_birth': forms.TextInput(attrs={'class': 'form-control', 'type': 'date'}),
             'address': forms.TextInput(attrs={'class': 'form-control', 'placeholder': _(u'Street address'), 'type': 'text'}),
             'zip_code': forms.TextInput(attrs={'class': 'form-control', 'placeholder': _(u'Postal code'), 'type': 'number'}),
-            'phone': forms.TextInput(attrs={'class': 'form-control', 'placeholder': _(u'Phone number'), 'type': 'tel'})}
+            'phone': forms.TextInput(attrs={'class': 'form-control', 'placeholder': _(u'Phone number'), 'type': 'tel'}),
+            'marketing_optin': forms.CheckboxInput()}
 
     def clean(self):
         cleaned_data = super(UserProfileForm, self).clean()
         if self.is_valid():
+
+            # Profile check must be consistent with RegisterForm
 
             # Check nick
             nick = cleaned_data['nick']
@@ -30,10 +33,14 @@ class UserProfileForm(forms.ModelForm):
                 self.add_error('nick', ugettext(u'Your desired nickname contains illegal characters. Valid: a-Z 0-9 - _'))
 
             # Check date of birth
-            # currently only checks that it is not after today
+            now = datetime.date.today()
             date = cleaned_data['date_of_birth']
-            if date >= datetime.date.today():
-                self.add_error('date_of_birth', ugettext(u'You seem to be from the future, please enter a more believable date of birth.'))
+            if date == now:
+                self.add_error('date_of_birth', ugettext(u'You seem to have been born today, that doesn\'t seem right.'))
+            if date >= now:
+                self.add_error('date_of_birth', ugettext(u'You seem to be from the future, that doesn\'t seem right.'))
+            if date < now.replace(year=(now.year - 150)):
+                self.add_error('date_of_birth', ugettext(u'You seem to be over 150 years old, that doesn\'t seem right.'))
 
             # ZIP code digits only
             zip_code = cleaned_data['zip_code']