+++ manage/clean-all.sh
@@ -21,5 +21,5 @@ rm -f studlan/settings/local.docker.py
 
 echo "Cleaning Docker ..."
 docker-compose -f setup/docker-compose.full.yml down
-docker-compose -f setup/docker-compose.test.yml down
+docker-compose -f setup/docker-compose.ci.yml down
 docker-compose -f setup/docker-compose.tiny.yml down